backend/requirements.txt
fastapi==0.129.0
uvicorn==0.41.0
pydantic==2.12.5
pydantic-core==2.41.5
PyJWT==2.12.1
argon2-cffi==25.1.0
passlib==1.7.4
SQLAlchemy==2.0.48
asyncpg==0.31.0
pytest==9.0.2
pytest-asyncio==1.3.0
httpx==0.28.1
pytest-cov==7.1.0
pytest-mock==3.15.1
azure-monitor-opentelemetry>=1.8.7,<2.0.0
python-dotenv==1.0.0
firebase-admin>=6.0.0,<7.0.0
google-generativeai>=0.3.0,<1.0.0
